+++ @pob/root/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.3](https://github.com/christophehurpeau/pob/compare/@pob/root@1.3.2...@pob/root@1.3.3) (2020-02-12)
+
+**Note:** Version bump only for package @pob/root
+
+
+
+
+
 ## [1.3.2](https://github.com/christophehurpeau/pob/compare/@pob/root@1.3.1...@pob/root@1.3.2) (2020-02-07)